refactor(library): Extract component to represent Mayoral elections
