+++ tests/test_historic_sopn_publish_dates.py
@@ -4,6 +4,7 @@ from uk_election_timetables.elections import (
     NorthernIrelandAssemblyElection,
     UKParliamentElection,
     PoliceAndCrimeCommissionerElection,
+    MayoralElection,
 )
 from uk_election_timetables.sopn import StatementPublishDate
 from datetime import timedelta, datetime
@@ -119,7 +120,9 @@ def test_police_and_crime_commissioner(row):
     ids=generate_test_id,
 )
 def test_mayoral(row):
-    expected_date = sopn_publish_date.mayor(read_date(row["election_date"]))
+    poll_date = read_date(row["election_date"])
+
+    expected_date = MayoralElection(poll_date).sopn_publish_date()
 
     actual_date = read_date(row["sopn_publish_date"])
